+++ NEWS.md
@@ -10,7 +10,8 @@
   
 **Improvements**
 
-* New geoms `geom_connect()`, `geom_connect_bars()` and `geom_prop_connect()`
+* New geoms `geom_connect()`, `geom_connect_bars()` and
+  `geom_prop_connect()` (#81)
 * New shortcut `auto_contrast` (#75)
 
 **Other changes**